Render index from server
