+++ package.json
@@ -12,6 +12,7 @@
   "license": "MIT",
   "dependencies": {
     "body-parser": "^1.20.2",
+    "ejs": "^3.1.9",
     "express": "^4.18.2",
     "fs-extra": "^11.2.0",
     "node-fetch": "^2.6.5"
