+++ public/another-test.html
@@ -1,0 +1,10 @@
+
+    <html>
+      <head></head>
+      <body>
+        <ul>
+          <li>Another test maaaaaan</li>
+        </ul>
+      </body>
+    </html>
+
