+++ public/from-server.html
@@ -1,0 +1,10 @@
+
+    <html>
+      <head></head>
+      <body>
+        <ul>
+          <li>FROM SERVER!</li>
+        </ul>
+      </body>
+    </html>
+
