+++ public/index.ejs
@@ -9,6 +9,8 @@
 </head>
 <body>
     <h1>Random Tiny Desk Concert</h1>
-    <main data-tiny-desk-concert></main>
+    <main data-tiny-desk-concert>
+      <a href='<%= tagline %>' target='blank'>Here's your random concert. Click to reveal!</a>
+    </main>
 </body>
 </html>
