+++ public/index.js
@@ -10,17 +10,5 @@ document.addEventListener('DOMContentLoaded', async (e) => {
 
           localStorage.setItem('concerts', JSON.stringify(concertList));
         }
-
-        const randIndex = getRandomInt(concertList.length - 1)
-        const randConcertLink = concertList[randIndex]
-
-        mainEl.innerHTML = `
-            <a href='${randConcertLink}' target='blank'>Here's your random concert. Click to reveal!</a>
-        `
     }
 })
-
-
-function getRandomInt(max) {
-    return Math.floor(Math.random() * max);
-}
