+++ public/test-build-command.html
@@ -1,0 +1,10 @@
+
+    <html>
+      <head></head>
+      <body>
+        <ul>
+          <li>Hi hello</li>
+        </ul>
+      </body>
+    </html>
+
