+++ public/will-this-work.html
@@ -1,0 +1,10 @@
+
+    <html>
+      <head></head>
+      <body>
+        <ul>
+          <li>Will this work in Vercel preview deployment?</li>
+        </ul>
+      </body>
+    </html>
+
