+++ views/index.ejs
@@ -1,0 +1,16 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Random Tiny Desk Concert</title>
+
+    <script src="index.js" type="module" defer></script>
+</head>
+<body>
+    <h1>Random Tiny Desk Concert</h1>
+    <main data-tiny-desk-concert>
+      <a href='<%= concertLink %>' target='blank'>Here's your random concert. Click to reveal!</a>
+    </main>
+</body>
+</html>
